Set desired capacity based on resource refresh or query.
The autoscale group name would already have been set if its a refresh call, as it is used to fetch the resource from the cloud. The only time it will be null is when the copyFrom is caled via a query.

desiredCapacity is set to null if not configured and is loaded as part of the refresh, but if it is loaded as part of a query then the actual value is used.