Make getData return lowercase value if present.
The attribute 'data' for resources (XssMatchTupple, SqlInjection, SizeConstraint, regexMatchTuple and ByteMatchTuple ) always gets saved as lowercase. So a conversion to lower if the config as mixed/upper case resolved the difference.

As the data attribute is also part of the primary key, this also resolves the issue where one of the above is created with a upper / mixed case 'data' attribute, and a corresponding gyro up tries to delete the one with teh lower case data-attribute and recreate again.