Fix S3 Domain Name getter.
The domain name getter searches for the region based on a bucket name. If it fails the getter should resolve to the defult domain name, rather than throwing an exception.

This handles error when creating a new bucket, where the getter would fail as the bucket does not exist when quering for the region.