Fix primary key null value if protocol set to -1.
When protocol is set to -1, It affects all the ports,  and specific 'from-port' and 'to-port' values cannot be set.

Changed the display string to show 'all' for 'protocol', 'from-port' and 'to-port' when protocol set to '-1'.

Add validation to check that 'from-port' and 'to-port' are not configured when 'protocol' set to '-1'.