Fix handling default egress rule.
1. Fix how to load/ignore default egress rule.

Previous implementation was only acounting for the protocol without looking at the the individual ip when ignoring the default rule, which would lead to ignoring all egress rules which had '-1' protocol.

2. Improve deleteing default rule.

Remove redundant check of 'from-port' and 'to-port' to be null when protocol is '-1' whenloading and checking default rules.

Add support for deleting ipv6 default rule.

Search the ipRange list for the default rule ip (0.0.0.0.0/0), rather than getting the  first one and expecting it to be the default.