Fix error updating subnet when availability zone not provided
Updating subnet throws an error if availability zone is not set and the updated subnet's availability zone is different than the previously configured subnet's availability zone.

The reason for this was that the availability zone was saved in state even if not configured.

This fixes it by only setting the availability zone if configured.