Refactor ELB instances to include state
Trying to refactor this to more closely match the API/Model of AWS while also providing the ability to get the current live instances within a Gyro workflow.

The issue is that if you use query for an ELB and grab the instances while in a workflow the instances might be missing since they could have been added in a previous workflow stage. Gyro doesn't automatically refresh after the stage so the next stage won't see them.

Specifically happens if an autoscale group is adding the instances to the ELB.