Fix gyro not picking up zip update
The earlier approach was saving the hash of the zip file in a separate field to pick up changes when teh path remains same.

This was before configured fields were the only fields whose diff was picked up.

The current approach appends the hash of the zip file to the path of the file, so any change in either the path or the content changes the field so gyro can pick up the change.