Fix refresh error when multiple versions are present
When multiple versions are there for a policy, the policy json should be fetched of the version that is active, in this case that is the version that is labeled default.

The previous approach was to use the last version id, which is not always the default one and can cause an issue.