Synchronize client creations
Aws Clients are created on the fly multi threaded and due to a potential race condition in the aws sdk client creation, sometimes the returned client object is null if simultaneous client creation of teh saem type is done.

While debugging it was seen that in the existing way, clients were created multiple times and some of them were null if the right race condition occured.

With a retry mechanism in place, if the client were null, we saw that some clients were repeatedly being created null same cause of teh race condition.

Having the client creation methods in a synchronized way mitigates this and makes sure that clients are created one at a time.

The same bedugging showed that all the unique clients were created only once.

For now this seems to be most logical step to take, and stops the error from reproducing.